+++ packages/queryable/behaviors/inject-headers.ts
@@ -1,4 +1,3 @@
-import { ensureHeaders } from "@pnp/core";
 import { Queryable2 } from "../queryable-2.js";
 
 export function InjectHeaders(headers: Record<string, string>): (instance: Queryable2) => Queryable2 {
@@ -7,7 +6,7 @@ export function InjectHeaders(headers: Record<string, string>): (instance: Query
 
         instance.on.pre(async function (url: string, init: RequestInit, result: any) {
 
-            init = ensureHeaders(init, headers);
+            init.headers = { ...init.headers, ...headers };
 
             return [url, init, result];
         });