+++ packages/sp/features/index.ts
@@ -1,11 +1,15 @@
 import "./site.js";
 import "./web.js";
 
+import { Features } from "./types.js";
+
 export {
     Feature,
     IFeature,
     Features,
-    IFeatures,
+    // ReturnType<Features> as IFeatures,
     IFeatureAddResult,
     IFeatureInfo,
 } from "./types.js";
+
+export type IFeatures = ReturnType<typeof Features>;