+++ packages/sp/webs/batching.ts
@@ -2,7 +2,7 @@ import { getGUID, isUrlAbsolute, combine, From_JulieHatesThisName } from "@pnp/c
 import { LogLevel } from "@pnp/logging";
 import { InjectHeaders, parseBinderWithErrorCheck, Queryable2 } from "@pnp/queryable";
 import { spPost } from "../operations";
-import { _SharePointQueryable } from "../sharepointqueryable";
+import { _SPQueryable } from "../sharepointqueryable";
 import { IWeb } from "./types.js";
 
 function BatchParse(): (instance: Queryable2) => Queryable2 {
@@ -14,7 +14,7 @@ function BatchParse(): (instance: Queryable2) => Queryable2 {
     });
 }
 
-class BatchQueryable extends _SharePointQueryable {
+class BatchQueryable extends _SPQueryable {
 
     constructor(web: IWeb, public requestBaseUrl = web.toUrl().replace(/_api\/.*$/i, "")) {
 
@@ -28,9 +28,6 @@ class BatchQueryable extends _SharePointQueryable {
     }
 }
 
-// TODO:: this needs to be reworked as a behavior meaning all requests would batch? How does that play? maybe need the creatBatch concept so you can get the execute
-// TODO: this would live on sp or web or site and get the url from there
-// TODO: how do we handle auth here? Inherit a batch queryable from the parent like "web" and clear out the other settings?
 // eslint-disable-next-line max-len
 export function createBatch(base: IWeb): [(instance: Queryable2) => Queryable2, () => Promise<void>] {